+++ packages/compile-common/test/shims.ts
@@ -1,15 +1,19 @@
 import assert from "assert";
-import { shims } from "../src";
+import { shims, Bytecode } from "../src";
+const { legacyToNew, newToLegacy } = shims;
 
-describe("shimBytecode", () => {
+describe("newToLegacy.shimBytecode", () => {
   it("handles undefined", () => {
-    assert.equal(shims.newToLegacy.shimBytecode(undefined), undefined);
+    assert.equal(newToLegacy.shimBytecode(undefined), undefined);
   });
 
   it("prepends 0x", () => {
     const bytes = "ffffffff";
 
-    assert.equal(shims.newToLegacy.shimBytecode({ bytes }), `0x${bytes}`);
+    assert.equal(
+      newToLegacy.shimBytecode({ bytes, linkReferences: [] }),
+      `0x${bytes}`
+    );
   });
 
   it("inlines an external link reference into underscores format", () => {
@@ -28,7 +32,7 @@ describe("shimBytecode", () => {
     //                  0 1 2 3 4 5 6 7 8 9
     const expected = "0x00__hello_________00";
 
-    assert.equal(shims.newToLegacy.shimBytecode(bytecode), expected);
+    assert.equal(newToLegacy.shimBytecode(bytecode), expected);
   });
 
   it("inlines a link reference with multiple offsets", () => {
@@ -47,7 +51,7 @@ describe("shimBytecode", () => {
     //                  0 1 2 3 4 5 6 7 8 9
     const expected = "0x__hi____00__hi____00";
 
-    assert.equal(shims.newToLegacy.shimBytecode(bytecode), expected);
+    assert.equal(newToLegacy.shimBytecode(bytecode), expected);
   });
 
   it("inlines two different link references", () => {
@@ -71,6 +75,61 @@ describe("shimBytecode", () => {
     //                  0 1 2 3 4 5 6 7 8 9
     const expected = "0x__hi____00__there_00";
 
-    assert.equal(shims.newToLegacy.shimBytecode(bytecode), expected);
+    assert.equal(newToLegacy.shimBytecode(bytecode), expected);
+  });
+});
+
+describe("legacyToNew.shimBytecode", () => {
+  it("removes 0x", function () {
+    const bytes = "ffffffff";
+    const expected = {
+      bytes,
+      linkReferences: []
+    } as Bytecode;
+
+    assert.deepEqual(legacyToNew.shimBytecode(`0x${bytes}`), expected);
+  });
+
+  it("externalizes a link reference in underscores format", function () {
+    //                  0 1 2 3 4 5 6 7 8 9
+    const bytecode = "0x00__hello_________00";
+
+    const expected = {
+      //      0 1 2 3 4 5 6 7 8 9
+      bytes: "00000000000000000000",
+      linkReferences: [
+        {
+          offsets: [1],
+          length: 8,
+          name: "hello"
+        }
+      ]
+    };
+
+    assert.deepEqual(legacyToNew.shimBytecode(bytecode), expected);
+  });
+
+  it("externalizes two different link references", function () {
+    //                  0 1 2 3 4 5 6 7 8 9
+    const bytecode = "0x__hi____00__there_00";
+
+    const expected = {
+      //      0 1 2 3 4 5 6 7 8 9
+      bytes: "00000000000000000000",
+      linkReferences: [
+        {
+          offsets: [0],
+          length: 4,
+          name: "hi"
+        },
+        {
+          offsets: [5],
+          length: 4,
+          name: "there"
+        }
+      ]
+    };
+
+    assert.deepEqual(legacyToNew.shimBytecode(bytecode), expected);
   });
 });
